+++ CHANGELOG.md
@@ -6,5 +6,8 @@
 - - Add `java.util.*` package to whitelist
 - - Implement `MinecraftClient.worldIsRainingAt` functionality
 
+**Fix:**
+- Fix Scrollbar in GUI screens shifting away when dragged with mouse
+
 **Download:**  
 You can download this release on [Modrinth](https://modrinth.com/mod/jcm) or [GitHub](https://github.com/DistrictOfJoban/Joban-Client-Mod/releases)